Made it so plugins that cannot be uninstalled are not added to the installed list when they are downloaded.
git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@4883 956a8c8d-885e-cb46-a0b9-a2f355b4c364

+++ AnomalousMedical/GUI/DownloadManagerGUI/ServerPluginDownloadInfo.cs
@@ -50,7 +50,14 @@ namespace Medical.GUI
             AtlasPlugin plugin = pluginDownload.Plugin;
             if (pluginDownload.LoadedSucessfully && plugin != null)
             {
-                return new UninstallInfo(plugin);
+                if (plugin.AllowUninstall)
+                {
+                    return new UninstallInfo(plugin);
+                }
+                else
+                {
+                    return null;
+                }
             }
             return new UpdateInfo(ImageKey, Name, String.Format("You must restart Anomalous Medical to finish installing {0}. You may install more things by selecting them and clicking install before restarting.", Name), ServerDownloadStatus.PendingInstall, false);
         }
